requirements.txt
# Data manipulation
pandas==2.1.0
numpy==1.24.3

# Visualization
matplotlib==3.7.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.0
pyspark==3.4.1
tensorflow==2.15.0
keras==2.15.0

# Utilities
joblib==1.3.2
jupyter==1.0.0
ipykernel==6.25.2

# Logging and configuration
pyyaml==6.0.1
python-dotenv==1.0.0